+++ src/consts/application.ts
@@ -1,13 +1,17 @@
-//External Links
+// External Links
 export const GITHUB_TOKEN_DOC = 'https://docs.github.com/en/github/authenticating-to-github/keeping-your-account-and-data-secure/creating-a-personal-access-token';
 
-//Stored keys
+// Stored keys
 export const FAVORITE_REPOS_KEY = 'favoritedRepos';
 export const ACCESS_TOKEN_SECRET = 'accessToken';
 
-//Views
+// View Item keys
+export const VIEW_ITEM_CATEGORY = 'categoryItem';
+export const VIEW_ITEM_REPOSITORY = 'repositoryItem';
+
+// Views
 export const GIT_MARKER_VIEW = 'gitmarkerView';
 
-//context
+// Context
 export const SET_CONTEXT = 'setContext';
 export const CONTEXT_CATEGORY_COUNT = 'categoryCount';