+++ src/consts/messages.ts
@@ -1,6 +1,8 @@
 // Messages
-export const TYPE_NAME_CATEGORY_MSG = 'Type a new name for';
+export const TYPE_NAME_CATEGORY_MSG = 'Type the category name';
+export const CHOOSE_CATEGORY = 'Please select the category for this bookmark';
 export const CATEGORY_ALREADY_EXISTS_MSG = 'already exists. Please try with another name.';
+export const ALREADY_EXISTS_MSG = ' already exists.';
 export const CLEAR_ALL_MSG = 'Are you sure you want to clear all categories?';
 export const SURE_REMOVING_MSG = 'Are you sure you want to remove';
 export const ERROR_FETCHING_DATA_MSG = 'Error fetching data from GitHub...';
@@ -18,3 +20,13 @@ export const NO_MSG = 'No';
 // PlaceHolders
 export const TYPE_SEARCH_TERM_PLACEHOLDER = 'Type a term to search on GitHub';
 export const TYPE_ACCESS_TOKEN_PLACEHOLDER = 'Type your GitHub personal access token';
+
+// UI
+export const TOOLTIP_OWNED_BY = ' 🧍  Owned by ';
+export const TOOLTIP_STARGAZERS = '⭐ ';
+export const TOOLTIP_STARGAZERS_LBL = ' stars';
+export const TOOLTIP_FORKS = ' 🍴  ';
+export const TOOLTIP_FORKS_LBL = ' forks';
+export const TOOLTIP_LICENSE = '📝 ';
+export const TOOLTIP_LANGUAGE = '🧬 Written in ';
+export const TOOLTIP_LASTSYNC = 'Last sync: ';