+++ src/services/tree-view-manager.ts
@@ -1,20 +1,30 @@
 import * as vscode from 'vscode';
 import { CONTEXT_CATEGORY_COUNT, GIT_MARKER_VIEW, SET_CONTEXT } from '../consts/application';
-import { REFRESH_REPOSITORY } from '../consts/commands';
-import { Category } from '../models/category';
+import { SYNC_REPOSITORY } from '../consts/commands';
+import { TOOLTIP_FORKS, TOOLTIP_FORKS_LBL, TOOLTIP_LANGUAGE, 
+			TOOLTIP_LASTSYNC, TOOLTIP_LICENSE, TOOLTIP_OWNED_BY, 
+			TOOLTIP_STARGAZERS, TOOLTIP_STARGAZERS_LBL } from '../consts/messages';
+import { CategoriesRepositories } from '../models/categories-repositories';
 import { GithubRepository } from '../models/github-repository';
 import { TreeDataItem } from '../models/tree-data-item';
 import { formatDate } from '../utils/datetime-helper';
 import { TreeDataItemProvider } from './tree-data-item-provider';
 
 export class TreeViewManager {
-	context: vscode.ExtensionContext;
 	dataProvider: TreeDataItemProvider;
-	categories: Category[] | undefined;
 	treeView: any;
 
-	constructor(context: vscode.ExtensionContext) {	
-		this.context = context;
+	private static _instance: TreeViewManager;
+
+	static init(): void {
+		TreeViewManager._instance = new TreeViewManager();
+	}
+	
+	static get instance(): TreeViewManager {
+		return TreeViewManager._instance;
+	}
+
+	constructor() {	
 		this.dataProvider = new TreeDataItemProvider();
 		this.treeView = vscode.window.createTreeView(GIT_MARKER_VIEW, {
 			showCollapseAll: true,
@@ -26,37 +36,54 @@ export class TreeViewManager {
 			this.click(e.selection));
 	}
 
-	refreshDataProvider(categories?: Category[]) {
-		if(categories) {			
+	buildDataProviderItems(categoriesRepositories: CategoriesRepositories) {
+		if(categoriesRepositories) {			
 			const dataItems: TreeDataItem[] = [];
-			categories.forEach(category => {
+
+			categoriesRepositories.categories.forEach(category => {
 				const categoryRepositories: TreeDataItem[] = [];
-				category.repositories.forEach(repository => {
-					let repositoryDataItem = new TreeDataItem(false, `${repository.name}`);
-					repositoryDataItem.customId = repository.id;
-					repositoryDataItem.parentId = category.id;
-					repositoryDataItem.url = repository.url;					
-					repositoryDataItem.tooltip = this.buildToolTip(repository);
-					repositoryDataItem.description = repository.stargazersCount > 0 
-						? `⭐${repository.stargazersCount}` 
-						: '';
+				category.repositories.forEach(repositoryId => {
+					const repository = categoriesRepositories
+						.repositories.filter(r=>r.id === repositoryId)[0];
 					
-					categoryRepositories.push(repositoryDataItem);
+					// Building repository dataItem
+					const repositoryDataItem = this
+						.buildRepositoryDataItem(category.id, repository);					
+					if(repositoryDataItem){
+						categoryRepositories.push(repositoryDataItem);
+					}										
 				});
 
+				// Building category dataItem
 				const categoryDataItem = new TreeDataItem(true, category.name, categoryRepositories);
 				categoryDataItem.id = category.id;
 				dataItems.push(categoryDataItem);
 			});
+			
+			vscode.commands.executeCommand(SET_CONTEXT, CONTEXT_CATEGORY_COUNT, 
+				categoriesRepositories.categories.length);
 
 			this.dataProvider.setTreeItems(dataItems);
 			this.dataProvider.refresh();
-
-			vscode.commands.executeCommand(SET_CONTEXT, CONTEXT_CATEGORY_COUNT, 
-				categories.length);
 		}
 	}
+	
+	private buildRepositoryDataItem(categoryId: string, repository: GithubRepository) : TreeDataItem | undefined {
+		let repositoryDataItem!: TreeDataItem;
+		if(repository) {
+			repositoryDataItem = new TreeDataItem(false, `${repository.name}`);
+			repositoryDataItem.customId = repository.id;
+			repositoryDataItem.parentId = categoryId;
+			repositoryDataItem.url = repository.url;					
+			repositoryDataItem.tooltip = this.buildToolTip(repository);
+			repositoryDataItem.description = repository.stargazersCount > 0 
+				? `⭐${repository.stargazersCount}` 
+				: '';
+		}
 
+		return repositoryDataItem;
+	}
+	
 	private buildToolTip(repository: GithubRepository): string {
 		const newLine = '\n';
 
@@ -67,20 +94,20 @@ export class TreeViewManager {
 			toolTipItems.push(`${repository.description}${newLine}`);
 		}
 
-		toolTipItems.push(` 🧍  Owned by ${repository.ownerName}`);
-		toolTipItems.push(`⭐ ${repository.stargazersCount} stars`);
-		toolTipItems.push(` 🍴  ${repository.forks} forks`);
+		toolTipItems.push(`${TOOLTIP_OWNED_BY}${repository.ownerName}`);
+		toolTipItems.push(`${TOOLTIP_STARGAZERS}${repository.stargazersCount}${TOOLTIP_STARGAZERS_LBL}`);
+		toolTipItems.push(`${TOOLTIP_FORKS}${repository.forks}${TOOLTIP_FORKS_LBL}`);
 
 		if(repository.language) { 
-			toolTipItems.push(`🧬 Written in ${repository.language}`);
+			toolTipItems.push(`${TOOLTIP_LANGUAGE}${repository.language}`);
 		} 
 			
 		if(repository.license) { 
-			toolTipItems.push(`📝 ${repository.license.name}`);	
+			toolTipItems.push(`${TOOLTIP_LICENSE}${repository.license.name}`);	
 		}
 
 		toolTipItems.push(`${newLine}`);
-		toolTipItems.push(`${formatDate(repository.lastSyncDate)}`);
+		toolTipItems.push(`${TOOLTIP_LASTSYNC}${formatDate(repository.lastSyncDate)}`);
 
 		return toolTipItems.join(newLine); 
 	}
@@ -89,7 +116,7 @@ export class TreeViewManager {
 		selected.forEach(element => {
 			if(element.url) {
 				vscode.env.openExternal(element.url);
-				vscode.commands.executeCommand(REFRESH_REPOSITORY, selected[0]);
+				vscode.commands.executeCommand(SYNC_REPOSITORY, selected[0]);
 			}
 		});
 	}