+++ CHANGELOG.md
@@ -5,11 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.1 Unreleased]
 
 - Initial release
-## [0.1.0] - 2021-08-14
 
+--------------------------------------------------------------------------------------------------------
+
+## [0.1.0] - 2021-08-14
 ### Added
 
 - Import bookmarks feature.
@@ -19,3 +21,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Better grouping for "..." view/title menu.
+
+--------------------------------------------------------------------------------------------------------
+
+## [1.0.0] - 2021-08-19
+### Added
+
+- If you sync a repository, it will do it for all its existing references within other categories.
+
+### Changed
+
+- Changed the way the data is organized and stored, resulting in a way smaller export file.
+- Centralized the search on GitHub at the top navigation menu for a better UI.
+
+--------------------------------------------------------------------------------------------------------